+++ test/application_system_test_case.rb
@@ -1,9 +1,10 @@
 require "test_helper"
+require_relative "../lib/chrome_for_testing"
 
 Selenium::WebDriver::Chrome::Service.driver_path = ENV["CHROMEDRIVER_PATH"] if ENV["CHROMEDRIVER_PATH"].present?
 
 class ApplicationSystemTestCase < ActionDispatch::SystemTestCase
   driven_by :selenium, using: :headless_chrome, screen_size: [ 1400, 1400 ] do |options|
-    options.binary = ENV["CHROME_BIN"] if ENV["CHROME_BIN"].present?
+    options.binary = ChromeForTesting.binary
   end
 end